update chain core test
